requirements.txt
google-api-python-client
flask
python-dotenv
diskcache
pre-commit
setuptools
pyyaml
markdown
icecream
